+++ src/main/java/net/enzo/plus/common/Config.java
@@ -7,6 +7,7 @@ import java.io.File;
 public class Config {
     public static boolean gottaGoFast = true;
     public static boolean craftOnly = false;
+    public static boolean boringChest = false;
 
     public static void letsConfigurate(File file) {
         Configuration conf = new Configuration(file);
@@ -15,7 +16,8 @@ public class Config {
             conf.load();
 
             gottaGoFast = conf.get("General", "Sonic Speed", gottaGoFast, "Set if Infinity armor gave you sonic speed").getBoolean(true);
-            craftOnly = conf.get("General", "Crafting Only", craftOnly, "Set if the game just accept craftings, for MineTweaker porpouses").getBoolean(false);
+            craftOnly = conf.get("General", "Crafting Only", craftOnly, "Set if the game just accept crafting, for MineTweaker porpoises").getBoolean(false);
+            boringChest = conf.get("General", "Boring Chest", boringChest, "Set if chest hold just 64 items per stack!").getBoolean(false);
         } catch (Exception e) {
             System.out.println("The reality is in collapse...");
             e.printStackTrace();