+++ src/main/java/net/enzo/plus/common/blocks/InfinityBlocks.java
@@ -1,6 +1,8 @@
 package net.enzo.plus.common.blocks;
 
 import cpw.mods.fml.common.registry.GameRegistry;
+import net.enzo.plus.common.tiles.TileAbsoluteCraftingTable;
+import net.enzo.plus.common.tiles.TileInfinityChest;
 import net.minecraft.block.Block;
 
 public class InfinityBlocks {
@@ -12,7 +14,12 @@ public class InfinityBlocks {
 
     public static void consumeUniverse() {
         absolute = GameRegistry.registerBlock(new BlockAbsoluteCraftingTable(), "absolute_craft");
-        //coke = GameRegistry.registerBlock(new BlockInfintiyChest(), "Infinity_Chest");
+        coke = GameRegistry.registerBlock(new BlockInfintiyChest(), "Infinity_Chest");
+        GameRegistry.registerTileEntity(TileAbsoluteCraftingTable.class, "Plus_Absolute_Crafting");
+        GameRegistry.registerTileEntity(TileInfinityChest.class, "Plus_Tile_Entity");
+        /*7377577274 bugs solved just registering the tile entities...
+        * WHY I'M SO STUPID, WHY GOD, ANSWER ME, I SWEAR, WHY I'M SO FUCKING DUMB?????
+        * */
 
         mixedBlock = GameRegistry.registerBlock(new BlockMixedBlock(), "mixed_block");
         //absolute = GameRegistry.registerBlock(new BlockStarCore(), "tile_star_core"); bruh