+++ src/main/java/net/enzo/plus/common/tiles/TileInfinityChest.java
@@ -1,102 +1,523 @@
 package net.enzo.plus.common.tiles;
 
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
+import net.enzo.plus.common.Config;
+import net.enzo.plus.common.blocks.BlockInfintiyChest;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockChest;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.inventory.ContainerChest;
 import net.minecraft.inventory.IInventory;
-import net.minecraft.inventory.ISidedInventory;
+import net.minecraft.inventory.InventoryLargeChest;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.nbt.NBTTagList;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.tileentity.TileEntityChest;
+import net.minecraft.util.AxisAlignedBB;
 
-public class TileInfinityChest extends TileInfinity implements IInventory, ISidedInventory {
-    private ItemStack[] stackInv = new ItemStack[108];
+import java.util.Iterator;
+import java.util.List;
 
-    @Override
-    public void readCustomNBT(NBTTagCompound tag) {
-        for (int i = 0; i < stackInv.length; ++i) {
-            stackInv[i] = ItemStack.loadItemStackFromNBT(tag.getCompoundTag("inv" + i));
-        }
+public class TileInfinityChest extends TileEntity implements IInventory {
+    private ItemStack[] chestContents = new ItemStack[117];
+    /** Determines if the check for adjacent chests has taken place. */
+    public boolean adjacentChestChecked;
+    /** Contains the chest tile located adjacent to this one (if any) */
+    public TileInfinityChest adjacentChestZNeg;
+    /** Contains the chest tile located adjacent to this one (if any) */
+    public TileInfinityChest adjacentChestXPos;
+    /** Contains the chest tile located adjacent to this one (if any) */
+    public TileInfinityChest adjacentChestXNeg;
+    /** Contains the chest tile located adjacent to this one (if any) */
+    public TileInfinityChest adjacentChestZPos;
+    /** The current angle of the lid (between 0 and 1) */
+    public float lidAngle;
+    /** The angle of the lid last tick */
+    public float prevLidAngle;
+    /** The number of players currently using this chest */
+    public int numPlayersUsing;
+    /** Server sync counter (once per 20 ticks) */
+    private int ticksSinceSync;
+    private int cachedChestType;
+    private String customName;
+    public TileInfinityChest() {
+        this.cachedChestType = -1;
+    }
+    
+    @SideOnly(Side.CLIENT)
+    public TileInfinityChest(int p_i2350_1_)
+    {
+        this.cachedChestType = p_i2350_1_;
     }
 
-    @Override
-    public void writeCustomNBT(NBTTagCompound tag) {
-        for (int x = 0; x < stackInv.length; ++x) {
-            NBTTagCompound craft = new NBTTagCompound();
-            stackInv[x].writeToNBT(craft);
-            tag.setTag("inv" + x, craft);
-        }
-    }
-
-    @Override
-    public int[] getAccessibleSlotsFromSide(int p_94128_1_) {
-        return new int[]{};
-    }
-
-    @Override
-    public boolean canInsertItem(int p_102007_1_, ItemStack p_102007_2_, int p_102007_3_) {
-        return false;
-    }
-
-    @Override
-    public boolean canExtractItem(int p_102008_1_, ItemStack p_102008_2_, int p_102008_3_) {
-        return false;
+    public void func_145976_a(String p_145976_1_)
+    {
+        this.customName = p_145976_1_;
     }
 
     @Override
     public int getSizeInventory() {
-        return 108;
+        return 117;
     }
 
     @Override
     public ItemStack getStackInSlot(int slot) {
-        return stackInv[slot];
+        return this.chestContents[slot];
     }
 
     @Override
     public ItemStack decrStackSize(int p_70298_1_, int p_70298_2_) {
-        return null;
+        if (this.chestContents[p_70298_1_] != null)
+        {
+            ItemStack itemstack;
+
+            if (this.chestContents[p_70298_1_].stackSize <= p_70298_2_)
+            {
+                itemstack = this.chestContents[p_70298_1_];
+                this.chestContents[p_70298_1_] = null;
+                this.markDirty();
+                return itemstack;
+            }
+            else
+            {
+                itemstack = this.chestContents[p_70298_1_].splitStack(p_70298_2_);
+
+                if (this.chestContents[p_70298_1_].stackSize == 0)
+                {
+                    this.chestContents[p_70298_1_] = null;
+                }
+
+                this.markDirty();
+                return itemstack;
+            }
+        }
+        else
+        {
+            return null;
+        }
     }
 
     @Override
     public ItemStack getStackInSlotOnClosing(int p_70304_1_) {
-        return null;
+        if (this.chestContents[p_70304_1_] != null)
+        {
+            ItemStack itemstack = this.chestContents[p_70304_1_];
+            this.chestContents[p_70304_1_] = null;
+            return itemstack;
+        }
+        else
+        {
+            return null;
+        }
     }
 
     @Override
     public void setInventorySlotContents(int p_70299_1_, ItemStack p_70299_2_) {
+        this.chestContents[p_70299_1_] = p_70299_2_;
 
+        if (p_70299_2_ != null && p_70299_2_.stackSize > this.getInventoryStackLimit())
+        {
+            p_70299_2_.stackSize = this.getInventoryStackLimit();
+        }
+
+        this.markDirty();
     }
 
     @Override
     public String getInventoryName() {
-        return null;
+        return this.hasCustomInventoryName() ? this.customName : "container.infinity";
     }
 
     @Override
     public boolean hasCustomInventoryName() {
-        return false;
+        return this.customName != null && this.customName.length() > 0;
+    }
+
+    public void setCustomName(String customName) {
+        this.customName = customName;
+    }
+
+    @Override
+    public void readFromNBT(NBTTagCompound p_145839_1_) {
+        super.readFromNBT(p_145839_1_);
+        NBTTagList nbttaglist = p_145839_1_.getTagList("ItemsInfinity", 10);
+        this.chestContents = new ItemStack[this.getSizeInventory()];
+
+        if (p_145839_1_.hasKey("CustomName", 8))
+        {
+            this.customName = p_145839_1_.getString("CustomName");
+        }
+
+        for (int i = 0; i < nbttaglist.tagCount(); ++i)
+        {
+            NBTTagCompound nbttagcompound1 = nbttaglist.getCompoundTagAt(i);
+            int j = nbttagcompound1.getByte("SlotInfinity") & 255;
+
+            if (j >= 0 && j < this.chestContents.length)
+            {
+                this.chestContents[j] = ItemStack.loadItemStackFromNBT(nbttagcompound1);
+            }
+        }
+    }
+    @Override
+    public void writeToNBT(NBTTagCompound p_145841_1_) {
+        super.writeToNBT(p_145841_1_);
+        NBTTagList nbttaglist = new NBTTagList();
+
+        for (int i = 0; i < this.chestContents.length; ++i)
+        {
+            if (this.chestContents[i] != null)
+            {
+                NBTTagCompound nbttagcompound1 = new NBTTagCompound();
+                nbttagcompound1.setByte("SlotInfinity", (byte)i);
+                this.chestContents[i].writeToNBT(nbttagcompound1);
+                nbttaglist.appendTag(nbttagcompound1);
+            }
+        }
+
+        p_145841_1_.setTag("ItemsInfinity", nbttaglist);
+
+        if (this.hasCustomInventoryName())
+        {
+            p_145841_1_.setString("CustomName", this.customName);
+        }
     }
 
     @Override
     public int getInventoryStackLimit() {
-        return Integer.MAX_VALUE; // Oh, yeah
+        return !Config.boringChest ? Integer.MAX_VALUE : 64; // Placeholder for future config feature
     }
 
     @Override
-    public boolean isUseableByPlayer(EntityPlayer player) {
-        return this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord) == this && player.getDistanceSq((double)this.xCoord + 0.5D, (double)this.yCoord + 0.5D, (double)this.zCoord + 0.5D) <= 64.0D;
+    public boolean isUseableByPlayer(EntityPlayer p_70300_1_) {
+        return this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord) != this ? false : p_70300_1_.getDistanceSq((double)this.xCoord + 0.5D, (double)this.yCoord + 0.5D, (double)this.zCoord + 0.5D) <= 64.0D;    
+    }
+
+    @Override
+    public void updateContainingBlockInfo() {
+        super.updateContainingBlockInfo();
+        this.adjacentChestChecked = false;
+    }
+    
+    private void func_145978_a(TileInfinityChest p_145978_1_, int p_145978_2_) {
+        if (p_145978_1_.isInvalid())
+        {
+            this.adjacentChestChecked = false;
+        }
+        else if (this.adjacentChestChecked)
+        {
+            switch (p_145978_2_)
+            {
+                case 0:
+                    if (this.adjacentChestZPos != p_145978_1_)
+                    {
+                        this.adjacentChestChecked = false;
+                    }
+
+                    break;
+                case 1:
+                    if (this.adjacentChestXNeg != p_145978_1_)
+                    {
+                        this.adjacentChestChecked = false;
+                    }
+
+                    break;
+                case 2:
+                    if (this.adjacentChestZNeg != p_145978_1_)
+                    {
+                        this.adjacentChestChecked = false;
+                    }
+
+                    break;
+                case 3:
+                    if (this.adjacentChestXPos != p_145978_1_)
+                    {
+                        this.adjacentChestChecked = false;
+                    }
+            }
+        }
+    }
+
+    public void setAdjacentChestChecked(boolean adjacentChestChecked) {
+        if (!this.adjacentChestChecked)
+        {
+            this.adjacentChestChecked = true;
+            this.adjacentChestZNeg = null;
+            this.adjacentChestXPos = null;
+            this.adjacentChestXNeg = null;
+            this.adjacentChestZPos = null;
+
+            if (this.func_145977_a(this.xCoord - 1, this.yCoord, this.zCoord))
+            {
+                this.adjacentChestXNeg = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord - 1, this.yCoord, this.zCoord);
+            }
+
+            if (this.func_145977_a(this.xCoord + 1, this.yCoord, this.zCoord))
+            {
+                this.adjacentChestXPos = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord + 1, this.yCoord, this.zCoord);
+            }
+
+            if (this.func_145977_a(this.xCoord, this.yCoord, this.zCoord - 1))
+            {
+                this.adjacentChestZNeg = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord - 1);
+            }
+
+            if (this.func_145977_a(this.xCoord, this.yCoord, this.zCoord + 1))
+            {
+                this.adjacentChestZPos = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord + 1);
+            }
+
+            if (this.adjacentChestZNeg != null)
+            {
+                this.adjacentChestZNeg.func_145978_a(this, 0);
+            }
+
+            if (this.adjacentChestZPos != null)
+            {
+                this.adjacentChestZPos.func_145978_a(this, 2);
+            }
+
+            if (this.adjacentChestXPos != null)
+            {
+                this.adjacentChestXPos.func_145978_a(this, 1);
+            }
+
+            if (this.adjacentChestXNeg != null)
+            {
+                this.adjacentChestXNeg.func_145978_a(this, 3);
+            }
+        }
+    }
+
+    public void checkForAdjacentChests()
+    {
+        if (!this.adjacentChestChecked)
+        {
+            this.adjacentChestChecked = true;
+            this.adjacentChestZNeg = null;
+            this.adjacentChestXPos = null;
+            this.adjacentChestXNeg = null;
+            this.adjacentChestZPos = null;
+
+            if (this.func_145977_a(this.xCoord - 1, this.yCoord, this.zCoord))
+            {
+                this.adjacentChestXNeg = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord - 1, this.yCoord, this.zCoord);
+            }
+
+            if (this.func_145977_a(this.xCoord + 1, this.yCoord, this.zCoord))
+            {
+                this.adjacentChestXPos = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord + 1, this.yCoord, this.zCoord);
+            }
+
+            if (this.func_145977_a(this.xCoord, this.yCoord, this.zCoord - 1))
+            {
+                this.adjacentChestZNeg = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord - 1);
+            }
+
+            if (this.func_145977_a(this.xCoord, this.yCoord, this.zCoord + 1))
+            {
+                this.adjacentChestZPos = (TileInfinityChest)this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord + 1);
+            }
+
+            if (this.adjacentChestZNeg != null)
+            {
+                this.adjacentChestZNeg.func_145978_a(this, 0);
+            }
+
+            if (this.adjacentChestZPos != null)
+            {
+                this.adjacentChestZPos.func_145978_a(this, 2);
+            }
+
+            if (this.adjacentChestXPos != null)
+            {
+                this.adjacentChestXPos.func_145978_a(this, 1);
+            }
+
+            if (this.adjacentChestXNeg != null)
+            {
+                this.adjacentChestXNeg.func_145978_a(this, 3);
+            }
+        }
+    }
+
+    private boolean func_145977_a(int p_145977_1_, int p_145977_2_, int p_145977_3_)
+    {
+        if (this.worldObj == null)
+        {
+            return false;
+        }
+        else
+        {
+            Block block = this.worldObj.getBlock(p_145977_1_, p_145977_2_, p_145977_3_);
+            return block instanceof BlockInfintiyChest && ((BlockInfintiyChest)block).field_149956_a == this.func_145980_j();
+        }
+    }
+    @Override
+    public void updateEntity()
+    {
+        super.updateEntity();
+        this.checkForAdjacentChests();
+        ++this.ticksSinceSync;
+        float f;
+
+        if (!this.worldObj.isRemote && this.numPlayersUsing != 0 && (this.ticksSinceSync + this.xCoord + this.yCoord + this.zCoord) % 200 == 0)
+        {
+            this.numPlayersUsing = 0;
+            f = 5.0F;
+            List list = this.worldObj.getEntitiesWithinAABB(EntityPlayer.class, AxisAlignedBB.getBoundingBox((double)((float)this.xCoord - f), (double)((float)this.yCoord - f), (double)((float)this.zCoord - f), (double)((float)(this.xCoord + 1) + f), (double)((float)(this.yCoord + 1) + f), (double)((float)(this.zCoord + 1) + f)));
+            Iterator iterator = list.iterator();
+
+            while (iterator.hasNext())
+            {
+                EntityPlayer entityplayer = (EntityPlayer)iterator.next();
+
+                if (entityplayer.openContainer instanceof ContainerChest)
+                {
+                    IInventory iinventory = ((ContainerChest)entityplayer.openContainer).getLowerChestInventory();
+
+                    if (iinventory == this || iinventory instanceof InventoryLargeChest && ((InventoryLargeChest)iinventory).isPartOfLargeChest(this))
+                    {
+                        ++this.numPlayersUsing;
+                    }
+                }
+            }
+        }
+
+        this.prevLidAngle = this.lidAngle;
+        f = 0.1F;
+        double d2;
+
+        if (this.numPlayersUsing > 0 && this.lidAngle == 0.0F && this.adjacentChestZNeg == null && this.adjacentChestXNeg == null)
+        {
+            double d1 = (double)this.xCoord + 0.5D;
+            d2 = (double)this.zCoord + 0.5D;
+
+            if (this.adjacentChestZPos != null)
+            {
+                d2 += 0.5D;
+            }
+
+            if (this.adjacentChestXPos != null)
+            {
+                d1 += 0.5D;
+            }
+
+            this.worldObj.playSoundEffect(d1, (double)this.yCoord + 0.5D, d2, "random.chestopen", 0.5F, this.worldObj.rand.nextFloat() * 0.1F + 0.9F);
+        }
+
+        if (this.numPlayersUsing == 0 && this.lidAngle > 0.0F || this.numPlayersUsing > 0 && this.lidAngle < 1.0F)
+        {
+            float f1 = this.lidAngle;
+
+            if (this.numPlayersUsing > 0)
+            {
+                this.lidAngle += f;
+            }
+            else
+            {
+                this.lidAngle -= f;
+            }
+
+            if (this.lidAngle > 1.0F)
+            {
+                this.lidAngle = 1.0F;
+            }
+
+            float f2 = 0.5F;
+
+            if (this.lidAngle < f2 && f1 >= f2 && this.adjacentChestZNeg == null && this.adjacentChestXNeg == null)
+            {
+                d2 = (double)this.xCoord + 0.5D;
+                double d0 = (double)this.zCoord + 0.5D;
+
+                if (this.adjacentChestZPos != null)
+                {
+                    d0 += 0.5D;
+                }
+
+                if (this.adjacentChestXPos != null)
+                {
+                    d2 += 0.5D;
+                }
+
+                this.worldObj.playSoundEffect(d2, (double)this.yCoord + 0.5D, d0, "random.chestclosed", 0.5F, this.worldObj.rand.nextFloat() * 0.1F + 0.9F);
+            }
+
+            if (this.lidAngle < 0.0F)
+            {
+                this.lidAngle = 0.0F;
+            }
+        }
+    }
+
+    /**
+     * Called when a client event is received with the event number and argument, see World.sendClientEvent
+     */
+    @Override
+    public boolean receiveClientEvent(int p_145842_1_, int p_145842_2_)
+    {
+        if (p_145842_1_ == 1)
+        {
+            this.numPlayersUsing = p_145842_2_;
+            return true;
+        }
+        else
+        {
+            return super.receiveClientEvent(p_145842_1_, p_145842_2_);
+        }
     }
 
     @Override
     public void openInventory() {
+        if (this.numPlayersUsing < 0)
+        {
+            this.numPlayersUsing = 0;
+        }
+
+        ++this.numPlayersUsing;
+        this.worldObj.addBlockEvent(this.xCoord, this.yCoord, this.zCoord, this.getBlockType(), 1, this.numPlayersUsing);
+        this.worldObj.notifyBlocksOfNeighborChange(this.xCoord, this.yCoord, this.zCoord, this.getBlockType());
+        this.worldObj.notifyBlocksOfNeighborChange(this.xCoord, this.yCoord - 1, this.zCoord, this.getBlockType());
 
     }
 
     @Override
     public void closeInventory() {
-
+        if (this.getBlockType() instanceof BlockInfintiyChest)
+        {
+            --this.numPlayersUsing;
+            this.worldObj.addBlockEvent(this.xCoord, this.yCoord, this.zCoord, this.getBlockType(), 1, this.numPlayersUsing);
+            this.worldObj.notifyBlocksOfNeighborChange(this.xCoord, this.yCoord, this.zCoord, this.getBlockType());
+            this.worldObj.notifyBlocksOfNeighborChange(this.xCoord, this.yCoord - 1, this.zCoord, this.getBlockType());
+        }
     }
 
     @Override
     public boolean isItemValidForSlot(int p_94041_1_, ItemStack p_94041_2_) {
-        return false;
+        return true;
+    }
+    public void invalidate()
+    {
+        super.invalidate();
+        this.updateContainingBlockInfo();
+        this.checkForAdjacentChests();
+    }
+
+    public int func_145980_j()
+    {
+        if (this.cachedChestType == -1)
+        {
+            if (this.worldObj == null || !(this.getBlockType() instanceof BlockInfintiyChest))
+            {
+                return 0;
+            }
+
+            this.cachedChestType = ((BlockInfintiyChest)this.getBlockType()).field_149956_a;
+        }
+
+        return this.cachedChestType;
     }
 }