+++ src/main/java/net/enzo/plus/common/tiles/TileInfinityChestRenderer.java
@@ -1,0 +1,158 @@
+package net.enzo.plus.common.tiles;
+
+import cpw.mods.fml.common.FMLLog;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
+import net.minecraft.block.Block;
+import net.minecraft.client.model.ModelChest;
+import net.minecraft.client.renderer.RenderBlocks;
+import net.minecraft.client.renderer.entity.RenderItem;
+import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.tileentity.TileEntityChest;
+import net.minecraft.util.ResourceLocation;
+import org.lwjgl.opengl.GL11;
+import org.lwjgl.opengl.GL12;
+import net.enzo.plus.common.blocks.BlockInfintiyChest;
+
+import java.util.Random;
+
+import static org.lwjgl.opengl.GL11.*;
+import static org.lwjgl.opengl.GL12.*;
+
+@SideOnly(Side.CLIENT)
+public class TileInfinityChestRenderer extends TileEntitySpecialRenderer {
+    ModelChest model = new ModelChest();
+    private ResourceLocation chestTexture = new ResourceLocation("plus:textures/blocks/infinity_chest_work.png");
+
+    @SuppressWarnings("unused")
+    private RenderBlocks renderBlocks;
+    private RenderItem itemRenderer;
+
+    private Random random;
+
+    public TileInfinityChestRenderer() {
+        this.model = new ModelChest();
+    }
+
+    public void render(TileInfinityChest tile, double x, double y, double z, float partialTick) {
+        
+    }
+    public void renderTileEntityAt(TileInfinityChest p_147500_1_, double p_147500_2_, double p_147500_4_, double p_147500_6_, float p_147500_8_) {
+        this.model = new ModelChest();
+        int i;
+
+        if (!p_147500_1_.hasWorldObj())
+        {
+            i = 0;
+        }
+        else
+        {
+            Block block = p_147500_1_.getBlockType();
+            i = p_147500_1_.getBlockMetadata();
+
+            if (block instanceof BlockInfintiyChest && i == 0)
+            {
+                try
+                {
+                    ((BlockInfintiyChest)block).func_149954_e(p_147500_1_.getWorldObj(), p_147500_1_.xCoord, p_147500_1_.yCoord, p_147500_1_.zCoord);
+                }
+                catch (ClassCastException e)
+                {
+                    FMLLog.severe("Attempted to render a chest at %d,  %d, %d that was not a chest", p_147500_1_.xCoord, p_147500_1_.yCoord, p_147500_1_.zCoord);
+                }
+                i = p_147500_1_.getBlockMetadata();
+            }
+
+            p_147500_1_.checkForAdjacentChests();
+        }
+
+        if (p_147500_1_.adjacentChestZNeg == null && p_147500_1_.adjacentChestXNeg == null)
+        {
+            ModelChest model = this.model;
+
+            if (p_147500_1_.adjacentChestXPos == null && p_147500_1_.adjacentChestZPos == null)
+            {
+                model = this.model;
+                this.bindTexture(chestTexture);
+            }
+
+            GL11.glPushMatrix();
+            GL11.glEnable(GL12.GL_RESCALE_NORMAL);
+            GL11.glColor4f(1.0F, 1.0F, 1.0F, 1.0F);
+            GL11.glTranslatef((float)p_147500_2_, (float)p_147500_4_ + 1.0F, (float)p_147500_6_ + 1.0F);
+            GL11.glScalef(1.0F, -1.0F, -1.0F);
+            GL11.glTranslatef(0.5F, 0.5F, 0.5F);
+            short short1 = 0;
+
+            if (i == 2)
+            {
+                short1 = 180;
+            }
+
+            if (i == 3)
+            {
+                short1 = 0;
+            }
+
+            if (i == 4)
+            {
+                short1 = 90;
+            }
+
+            if (i == 5)
+            {
+                short1 = -90;
+            }
+
+            if (i == 2 && p_147500_1_.adjacentChestXPos != null)
+            {
+                GL11.glTranslatef(1.0F, 0.0F, 0.0F);
+            }
+
+            if (i == 5 && p_147500_1_.adjacentChestZPos != null)
+            {
+                GL11.glTranslatef(0.0F, 0.0F, -1.0F);
+            }
+
+            GL11.glRotatef((float)short1, 0.0F, 1.0F, 0.0F);
+            GL11.glTranslatef(-0.5F, -0.5F, -0.5F);
+            float f1 = p_147500_1_.prevLidAngle + (p_147500_1_.lidAngle - p_147500_1_.prevLidAngle) * p_147500_8_;
+            float f2;
+
+            if (p_147500_1_.adjacentChestZNeg != null)
+            {
+                f2 = p_147500_1_.adjacentChestZNeg.prevLidAngle + (p_147500_1_.adjacentChestZNeg.lidAngle - p_147500_1_.adjacentChestZNeg.prevLidAngle) * p_147500_8_;
+
+                if (f2 > f1)
+                {
+                    f1 = f2;
+                }
+            }
+
+            if (p_147500_1_.adjacentChestXNeg != null)
+            {
+                f2 = p_147500_1_.adjacentChestXNeg.prevLidAngle + (p_147500_1_.adjacentChestXNeg.lidAngle - p_147500_1_.adjacentChestXNeg.prevLidAngle) * p_147500_8_;
+
+                if (f2 > f1)
+                {
+                    f1 = f2;
+                }
+            }
+
+            f1 = 1.0F - f1;
+            f1 = 1.0F - f1 * f1 * f1;
+            model.chestLid.rotateAngleX = -(f1 * (float)Math.PI / 2.0F);
+            model.renderAll();
+            GL11.glDisable(GL12.GL_RESCALE_NORMAL);
+            GL11.glPopMatrix();
+            GL11.glColor4f(1.0F, 1.0F, 1.0F, 1.0F);
+        }
+    }
+
+
+    @Override
+    public void renderTileEntityAt(TileEntity p_147500_1_, double p_147500_2_, double p_147500_4_, double p_147500_6_, float p_147500_8_) {
+        this.renderTileEntityAt((TileInfinityChest) p_147500_1_, p_147500_2_, p_147500_4_, p_147500_6_, p_147500_8_);
+    }
+}